requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
pydantic>=2.10.0
google-genai>=1.5.0
google-adk>=0.3.0
google-cloud-storage>=2.19.0
google-cloud-aiplatform>=1.80.0
Pillow>=11.1.0
aiofiles>=24.1.0
websockets>=14.2
